// src/constants/enums.ts
export enum SplitStrategy {
  MIXED_SENTENCES = "mixed_sentences",
  ULTRADENSE = "ultradense",
  WORD = "word",
  HYPERDENSE = "hyperdense",
  SHORT_SENTENCE = "short_sentence",
  HIGHPACED = "highpaced",
  DRAMATIC = "dramatic",
  PARAGRAPH = "paragraph",
  EPIC = "epic",
  CINEMATIC = "cinematic",
  ARTISTIC = "artistic",
  SLOWPACED = "slowpaced",
  SUPER_SLOW = "super_slow",
  SENTENCE = "sentence",
  DOCUMENTARY = "documentary",
  LONG_SENTENCE = "long_sentence",
}